requirements.txt
textgrad 
python-dotenv